+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [#4] - added domain_enhancer import
 - [#2] - transferred img_enhance function to domain_enhancer.py, added gif_enhance, changed input to parser
 - [#1] - update folder structure
 - [#0] - init repo